// booking-personal-trainer/src/common/enums/billing/billing.enum.ts
export enum BillableTargetType {
  WORKOUT = 'WORKOUT',
}

export enum BillingChargeStatus {
  DRAFT = 'DRAFT',
  ACTIVE = 'ACTIVE',
  EXPIRED = 'EXPIRED',
  CANCELLED = 'CANCELLED',
  VOID = 'VOID',
  CONVERTED = 'CONVERTED',
}

export enum PaymentStatus {
  NONE = 'NONE',
  REQUIRES_PAYMENT_METHOD = 'REQUIRES_PAYMENT_METHOD',
  REQUIRES_CONFIRMATION = 'REQUIRES_CONFIRMATION',
  REQUIRES_ACTION = 'REQUIRES_ACTION',
  PROCESSING = 'PROCESSING',
  PAID = 'PAID',
  FAILED = 'FAILED',
  CANCELLED = 'CANCELLED',
  REFUNDED = 'REFUNDED',
}